test(drop-vector-index): dedicated e2e package with lifecycle and write-matrix coverage
